pyproject.toml
[project]
name = "agentsphere"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "2captcha-python>=1.5.1",
    "anthropic[vertex]>=0.49.0",
    "bs4>=0.0.2",
    "duckduckgo-search>=7.5.4",
    "fastapi[standard]>=0.115.12",
    "free-proxy>=1.1.3",
    "gitpython>=3.1.44",
    "google-cloud-firestore>=2.20.1",
    "google-genai>=1.9.0",
    "litellm>=1.63.14",
    "markdownify>=1.1.0",
    "mongomock>=4.3.0",
    "ollama>=0.4.7",
    "passlib[bcrypt]>=1.7.4",
    "playwright>=1.51.0",
    "pydantic-settings>=2.8.1",
    "pylint>=3.3.6",
    "pymilvus>=2.5.6",
    "pymongo>=4.11.3",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "requests>=2.32.3",
    "responses>=0.25.7",
    "selenium>=4.30.0",
    "selenium-stealth>=1.0.6",
    "sse-starlette>=2.2.1",
    "uvicorn>=0.34.0",
    "webdriver-manager>=4.0.2",
    "websockets>=15.0.1",
]
